+++ testing-android/src/androidTest/java/com/yubico/yubikit/testing/sd/SecurityDomainTests.java
@@ -22,7 +22,7 @@ import org.junit.runners.Suite;
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
         Scp03Tests.class,
-        Scp11aTests.class
+        Scp11Tests.class
 })
 public class SecurityDomainTests {
 }
